Project successfuly-ended.

+++ app/Http/Controllers/IdeaController.php
@@ -4,12 +4,15 @@ declare(strict_types=1);
 
 namespace App\Http\Controllers;
 
+use App\Actions\CreateIdea;
+use App\Actions\UpdateIdea;
 use App\Http\Requests\StoreIdeaRequest;
 use App\Http\Requests\UpdateIdeaRequest;
 use App\IdeaStatus;
 use App\Models\Idea;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class IdeaController extends Controller
 {
@@ -46,22 +49,10 @@ class IdeaController extends Controller
     /**
      * Store a newly created resource in storage.
      */
-    public function store(StoreIdeaRequest $request)
+    public function store(StoreIdeaRequest $request, CreateIdea $action)
     {
-        // dd($request->all());
 
-        $idea = Auth::user()->ideas()->create($request->safe()->except(['steps', 'image']));
-
-        $idea->steps()->createMany(
-            collect($request->steps)->map(fn ($step) => ['description' => $step])
-        );
-        if ($request->hasFile('image')) {
-            $imagePath = $request->image->store('ideas', 'public');
-
-            $idea->update([
-                'image_path' => $imagePath,
-            ]);
-        }
+        $action->handle($request->validated());
 
         return to_route('idea.index')->with('succes', 'Idea created!');
 
@@ -72,6 +63,8 @@ class IdeaController extends Controller
      */
     public function show(Idea $idea)
     {
+        Gate::authorize('workWithIdea', $idea);
+
         return view('idea.show', [
             'idea' => $idea,
         ]);
@@ -82,15 +75,20 @@ class IdeaController extends Controller
      */
     public function edit(Idea $idea): void
     {
-        //
+        Gate::authorize('workWithIdea', $idea);
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(UpdateIdeaRequest $request, Idea $idea): void
+    public function update(UpdateIdeaRequest $request, Idea $idea, UpdateIdea $action)
     {
-        //
+
+        Gate::authorize('workWithIdea', $idea);
+
+        $action->handle($request->validated(), $idea);
+
+        return to_route('idea.show', $idea)->with('succes', 'Idea updated!');
     }
 
     /**
@@ -98,7 +96,7 @@ class IdeaController extends Controller
      */
     public function destroy(Idea $idea)
     {
-        // authorize that this is alowed
+        Gate::authorize('workWithIdea', $idea);
         $idea->delete();
 
         return to_route('idea.index');
